+++ Clases/08_MongoDB/06_Express_and_Mongoose_II/controllers/user.controller.js
@@ -2,21 +2,31 @@
 // CONTROLLER SETUP - User
 // ---------------------------------------------------
 
-// 1) Importing Model
+// 1) Importing External Libraries
+const mongoose = require('mongoose');
+const { ObjectId } = mongoose.Types; // Destructuring assignment to get ObjectId
+
+// 2) Importing Model
 const UserModel = require("../models/user.model");
 
-// 2) Exporting Controller functions
+// 3) Exporting Controller functions
 module.exports = {
   // 2.1) READ METHODS
   getAllUsers: (req, res) => {
-    UserModel.find({}).populate("pets")
+    UserModel.find({})
+      .populate("pets")
       .then((allUsers) => res.status(200).json(allUsers))
       .catch((err) =>
         res.status(500).json({ message: "Something went wrong", error: err })
       );
   },
   getOneUserById: (req, res) => {
-    UserModel.findOne({ _id: req.params.id }).populate("pets")
+    if (!ObjectId.isValid(req.params.id))
+      return res
+        .status(400)
+        .json({ message: "UUID doesn't match the specified format" });
+    UserModel.findOne({ _id: req.params.id })
+      .populate("pets")
       .then((oneSingleUser) => {
         if (oneSingleUser) {
           res.status(200).json(oneSingleUser);
@@ -29,7 +39,8 @@ module.exports = {
       );
   },
   findOneUserByEmail: (req, res) => {
-    UserModel.findOne({ email: req.query.email }).populate("pets")
+    UserModel.findOne({ email: req.query.email })
+      .populate("pets")
       .then((oneSingleUser) => {
         if (oneSingleUser) {
           res.status(200).json(oneSingleUser);
@@ -51,12 +62,17 @@ module.exports = {
       );
   },
   // 2.3) UPDATE METHODS
-  updateExisitingUser: (req, res) => {
+  updateOneUserById: (req, res) => {
+    if (!ObjectId.isValid(req.params.id))
+      return res
+        .status(400)
+        .json({ message: "UUID doesn't match the specified format" });
     const updateOptions = {
       new: true, // Return the updated document
       runValidators: true, // Enforce validation during update
     };
-    UserModel.findOneAndUpdate({ _id: req.params.id }, req.body, updateOptions).populate("pets")
+    UserModel.findOneAndUpdate({ _id: req.params.id }, req.body, updateOptions)
+      .populate("pets")
       .then((updatedUser) => {
         if (updatedUser) {
           res.status(200).json(updatedUser);
@@ -78,6 +94,10 @@ module.exports = {
       );
   },
   deleteOneUserById: (req, res) => {
+    if (!ObjectId.isValid(req.params.id))
+      return res
+        .status(400)
+        .json({ message: "UUID doesn't match the specified format" });
     UserModel.deleteOne({ _id: req.params.id })
       .then((result) => {
         if (result.deletedCount > 0) {
